Add server test
