+++ package/tests/conftest.py
@@ -104,3 +104,133 @@ def example_catalog():
         },
         layers={"raw": {"raw_data",}, "model_inputs": {"model_inputs",}},
     )
+
+
+def assert_nodes_equal(response_nodes, expected_nodes):
+    node_sort_key = lambda x: x["id"]
+    for response_node, expected_node in zip(
+        sorted(response_nodes, key=node_sort_key),
+        sorted(expected_nodes, key=node_sort_key),
+    ):
+        response_node_tags = response_node.pop("tags")
+        expected_node_tags = expected_node.pop("tags")
+        assert sorted(response_node_tags) == sorted(expected_node_tags)
+        assert response_node == expected_node
+
+
+def assert_edges_equal(response_edges, expected_edges):
+    edge_sort_key = lambda x: x["source"]
+    assert sorted(response_edges, key=edge_sort_key) == sorted(
+        expected_edges, key=edge_sort_key
+    )
+
+
+@pytest.fixture
+def assert_example_data():
+    """Assert a json response against the example pipelines data above"""
+
+    def _assert_graph_response(response_data):
+        """Assert graph response for the `example_pipelines` and `example_catalog`
+        fixtures."""
+        expected_edges = [
+            {"source": "7b140b3f", "target": "d5a8b994"},
+            {"source": "56118ad8", "target": "0ecea0de"},
+            {"source": "13399a82", "target": "56118ad8"},
+            {"source": "f1f1425b", "target": "7b140b3f"},
+            {"source": "0ecea0de", "target": "7b140b3f"},
+            {"source": "c506f374", "target": "56118ad8"},
+        ]
+        assert_edges_equal(response_data.pop("edges"), expected_edges)
+        # compare nodes
+        expected_nodes = [
+            {
+                "id": "56118ad8",
+                "name": "Process Data",
+                "full_name": "process_data",
+                "tags": ["split"],
+                "pipelines": ["__default__", "data_processing"],
+                "modular_pipelines": ["uk", "uk.data_processing"],
+                "type": "task",
+                "parameters": {"train_test_split": 0.1},
+            },
+            {
+                "id": "13399a82",
+                "name": "Uk.data Processing.raw Data",
+                "full_name": "uk.data_processing.raw_data",
+                "tags": ["split"],
+                "pipelines": ["__default__", "data_processing"],
+                "modular_pipelines": ["uk", "uk.data_processing"],
+                "type": "data",
+                "layer": None,
+            },
+            {
+                "id": "c506f374",
+                "name": "Params:train Test Split",
+                "full_name": "params:train_test_split",
+                "tags": ["split"],
+                "pipelines": ["__default__", "data_processing"],
+                "modular_pipelines": [],
+                "type": "parameters",
+                "layer": None,
+            },
+            {
+                "id": "0ecea0de",
+                "name": "Model Inputs",
+                "full_name": "model_inputs",
+                "tags": ["train", "split"],
+                "pipelines": ["__default__", "data_science", "data_processing"],
+                "modular_pipelines": [],
+                "type": "data",
+                "layer": "model_inputs",
+            },
+            {
+                "id": "7b140b3f",
+                "name": "Train Model",
+                "full_name": "train_model",
+                "tags": ["train"],
+                "pipelines": ["__default__", "data_science"],
+                "modular_pipelines": ["uk", "uk.data_science"],
+                "type": "task",
+                "parameters": {"train_test_split": 0.1, "num_epochs": 1000},
+            },
+            {
+                "id": "f1f1425b",
+                "name": "Parameters",
+                "full_name": "parameters",
+                "tags": ["train"],
+                "pipelines": ["__default__", "data_science"],
+                "modular_pipelines": [],
+                "type": "parameters",
+                "layer": None,
+            },
+            {
+                "id": "d5a8b994",
+                "name": "Uk.data Science.model",
+                "full_name": "uk.data_science.model",
+                "tags": ["train"],
+                "pipelines": ["__default__", "data_science"],
+                "modular_pipelines": ["uk", "uk.data_science"],
+                "type": "data",
+                "layer": None,
+            },
+        ]
+        assert_nodes_equal(response_data.pop("nodes"), expected_nodes)
+
+        # compare the rest
+        assert response_data == {
+            "tags": ["split", "train"],
+            "layers": [],
+            "pipelines": [
+                {"id": "__default__", "name": "Default"},
+                {"id": "data_science", "name": "Data Science"},
+                {"id": "data_processing", "name": "Data Processing"},
+            ],
+            "modular_pipelines": [
+                {"id": "uk", "name": "Uk"},
+                {"id": "uk.data_processing", "name": "Data Processing"},
+                {"id": "uk.data_science", "name": "Data Science"},
+            ],
+            "selected_pipeline": "__default__",
+        }
+
+    yield _assert_graph_response
